+++ ui/litellm-dashboard/src/app/(dashboard)/access-groups/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { AccessGroupsPage } from "@/components/AccessGroups/AccessGroupsPage";
+import { AccessGroupsPage } from "./components/AccessGroupsPage";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
 
 export default function AccessGroups() {
